Added inactive issues checking
